requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.32.0
frozenlist==1.3.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pycodestyle==2.8.0
pyopenvdb-3.8==0.1.5
pyparsing==3.0.8
python-dateutil==2.8.2
scipy==1.8.0
six==1.16.0
toml==0.10.2
vtk==9.1.0
wslink==1.5.3
yarl==1.7.2
